Move from MCPanel to native split view
